+++ src/Api/Dfe.Complete.Api/Controllers/TasksDataController.cs
@@ -228,5 +228,24 @@ namespace Dfe.Complete.Api.Controllers
             await sender.Send(request, cancellationToken);
             return NoContent();
         }
+        /// <summary>
+        /// Confirm the date the academy opened for conversion project.
+        /// </summary>
+        /// <param name="request">The update command.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPatch]
+        [Route("TaskData/ConfirmAcademyOpenedDate")]
+        [SwaggerResponse(204, "Confirm the date the academy opened for conversion project successfully.")]
+        [SwaggerResponse(400, "Invalid request data.")]
+        [SwaggerResponse(404, "Project not found.")]
+        public async Task<IActionResult> UpdateConfirmAcademyOpenedDateTaskAsync(
+            [FromBody] UpdateConfirmAcademyOpenedDateTaskCommand request,
+            CancellationToken cancellationToken)
+        {
+            await sender.Send(request, cancellationToken);
+            return NoContent();
+        }
+        
     }
 }