+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/ArticlesOfAssociationTask/ArticlesOfAssociationTask.cshtml.cs
@@ -3,7 +3,6 @@ using Dfe.Complete.Constants;
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.ValueObjects;
 using Dfe.Complete.Extensions;
-using Dfe.Complete.Models;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -60,7 +59,7 @@ namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.ArticlesOfAssociationTask
         public async Task<IActionResult> OnPost()
         {
             await sender.Send(new UpdateArticleOfAssociationTaskCommand(new TaskDataId(TasksDataId.GetValueOrDefault())!, Type, NotApplicable, Cleared, Received, Sent, Signed, Saved));
-            TempData.SetNotification(NotificationType.Success, "Success", "Task updated successfully");
+            TempData.SetTaskSuccessNotification();
             return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
         }
     }