+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/HandoverWithDeliveryOfficerTask/HandoverWithDeliveryOfficerTask.cshtml.cs
@@ -2,8 +2,7 @@ using Dfe.Complete.Application.Projects.Commands.TaskData;
 using Dfe.Complete.Constants; 
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.ValueObjects;
-using Dfe.Complete.Extensions;
-using Dfe.Complete.Models;
+using Dfe.Complete.Extensions; 
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -55,7 +54,7 @@ public class HandoverWithDeliveryOfficerTaskModel(ISender sender, IAuthorization
     public async Task<IActionResult> OnPost()
     {
         await sender.Send(new UpdateHandoverWithDeliveryOfficerTaskCommand(new TaskDataId(TasksDataId.GetValueOrDefault())!, Type, NotApplicable, ReviewProjectInformation, MakeNotes, AttendHandoverMeeting));
-        TempData.SetNotification(NotificationType.Success, "Success", "Task updated successfully");
+        TempData.SetTaskSuccessNotification();
         return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
     }
 }